#1308 Delete app/ui/screen_controllers/ (post-#1287 cleanup, closes #1287/#1193 OoS-B)
Every screen migrated to the entity-driven UI path (#1290 Paused pilot ->
#1305 Gameplay HUD); ui_render_system::render_ui_entities and ui_update_system
fully own runtime UI. The legacy OOP bridge in app/ui/screen_controllers/
and the matching *_layout.h god-struct headers are now dead code.

Deletions
- app/ui/screen_controllers/ (8 controller pairs + screen_controller_base.h,
  ~481 LOC).
- app/ui/generated/{game_over,paused,settings,song_complete,title,tutorial}_layout.h
  (6 god-struct headers).
- tools/rguilayout/SUMMARY.md (already self-marked SUPERSEDED; described the
  deleted manual-extraction workflow).

Required moves
- tutorial_screen_continue() relocates from the deleted controller into
  app/systems/ui_update_system.cpp at file scope, declared in
  ui_update_system.h. Only consumers were the ui_update_system action
  handler and the FTUE end-to-end test; both updated to the new location.
- RAYGUI_IMPLEMENTATION ownership moves from title_screen_controller.cpp
  to a new 1-line TU app/ui/raygui_impl.cpp. CMake gates the macro on this
  file alone and excludes it from unity builds (replaces the parallel
  set_source_files_properties block from the deleted controller). The
  earlier list(FILTER UI_SOURCES EXCLUDE ... raygui_impl.cpp) is dropped so
  the file participates in UI_SOURCES.

Test updates
- tests/test_game_state_extended.cpp: include path changes from
  ui/screen_controllers/tutorial_screen_controller.h to
  systems/ui_update_system.h.
- tests/test_helpers_and_functions.cpp: drop screen_controller_base.h
  include and the DummyScreenController test (screen_controller<> helper
  is gone).
- tests/test_phase_transition_canonical.cpp: update the fictional source
  fixture path to app/systems/ui_update_system.cpp; rewrite the entry-input
  guard test to assert UI_ENTRY_DEBOUNCE / GAME_OVER_INPUT_DELAY /
  SONG_COMPLETE_INPUT_DELAY all appear in ui_update_system.cpp (which is
  where the effective_debounce table now lives).

Doc cleanup
- tools/rguilayout/INTEGRATION.md: replace the 'Coexistence with the legacy
  *_layout.h pipeline' section (no coexistence remains) with the live
  runtime pipeline description pointing at screen_lifecycle_system /
  ui_render_system / ui_update_system.
- Inline comments referencing the legacy controller path in
  screen_lifecycle_system.h, screen_spawners.h, ui_update_system.{h,cpp},
  and ui_render_system.cpp updated to reflect post-deletion reality.

Verification
- cmake --build build -j clean under -Wall -Wextra -Werror on macOS.
- ./build/shapeshifter_tests: 933/933 passing (was 934; -1 from removing
  the obsolete screen_controller helper test).
- python3 -m unittest tools.test_rguilayout_codegen: 13/13 pass.
- python3 tools/check_enum_allowlist.py: clean.
- nm libshapeshifter_lib.a: GuiButton symbol defined exactly once (raygui
  ODR preserved).
- switch count in app/: 10, unchanged from baseline (#1309 work deferred).

After this lands, the last remaining done-criterion on #1287 and on #1193
('Delete app/ui/screen_controllers/') flips, closing both umbrella issues.
