--- app/ui/screen_controllers/paused_screen_controller.h
@@ -1,6 +1,0 @@
-#pragma once
-
-#include <entt/entt.hpp>
-
-void init_paused_screen_ui();
-void render_paused_screen_ui(entt::registry& reg);
