--- app/ui/screen_controllers/settings_screen_controller.h
@@ -1,6 +1,0 @@
-#pragma once
-
-#include <entt/entt.hpp>
-
-void init_settings_screen_ui();
-void render_settings_screen_ui(entt::registry& reg);
