--- app/ui/screen_controllers/song_complete_screen_controller.h
@@ -1,6 +1,0 @@
-#pragma once
-
-#include <entt/entt.hpp>
-
-void init_song_complete_screen_ui();
-void render_song_complete_screen_ui(entt::registry& reg);
