--- app/ui/screen_controllers/title_screen_controller.h
@@ -1,6 +1,0 @@
-#pragma once
-
-#include <entt/entt.hpp>
-
-void init_title_screen_ui();
-void render_title_screen_ui(entt::registry& reg);
