--- app/ui/screen_controllers/tutorial_screen_controller.h
@@ -1,7 +1,0 @@
-#pragma once
-
-#include <entt/entt.hpp>
-
-void tutorial_screen_continue(entt::registry& reg);
-void init_tutorial_screen_ui();
-void render_tutorial_screen_ui(entt::registry& reg);
